+++ Tests/Functional/Configuration/PageFinderTest.php
@@ -30,14 +30,13 @@ class PageFinderTest extends FunctionalTestCase
     {
         parent::setUp();
         $this->testingFramework = new \Tx_Oelib_TestingFramework('tx_oelib');
-        $this->testingFramework->setResetAutoIncrementThreshold(99999999);
 
         $this->subject = \Tx_Oelib_PageFinder::getInstance();
     }
 
     protected function tearDown()
     {
-        $this->testingFramework->cleanUp();
+        $this->testingFramework->cleanUpWithoutDatabase();
         parent::tearDown();
     }